requirements.txt
anthropic
python-dotenv
httpx
matplotlib
numpy
